Build the initial sender view.
Forms the basis of the sender application by showing the available
albums. This makes use of the Flickr JS wrapper, and adds a second
wrapper on top of this to handle the awkward design.

I'll tidy this up before the talk if I get time, but the current
focus is just making something work as quickly as possible.